pyproject.toml
[tool.poetry]
name = "climate"
version = "0.1.0"
description = "Command line arguments dependency injection."
authors = ["Oliver Berger <oliver@digitalarchitekt.de>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "*"
typing_inspect = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-randomly = "*"
pytest-watch = "*"
pytest-cache = "*"
pytest-mock = "*"
pytest-cov = "*"
mock = "*"
ipython = "*"
pdbpp = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
